# kotlin/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: Example spec of the well known JSON Placeholder website
  title: JSON Placeholder OpenAPI
  version: 0.1.15
servers:
- url: https://jsonplaceholder.typicode.com
tags:
- name: qr
paths:
  /code/qr/edit/{serialNumber}:
    post:
      operationId: editQRCode
      parameters:
      - explode: false
        in: path
        name: serialNumber
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QRCodeCreateData'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeBaseData'
          description: Successfully edited QR Code
      summary: Edit QR Code and returns Base Data
      tags:
      - qr
  /code/qr/create:
    post:
      operationId: createQRCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QRCodeCreateData'
        description: QR Code Data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeBaseData'
          description: Successfuly created QR Code
      summary: Creates a new QR Code and returns Base Data
      tags:
      - qr
  /code/qr/{serialNumber}:
    get:
      operationId: getQRCode
      parameters:
      - explode: false
        in: path
        name: serialNumber
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QRCodeBaseData'
          description: Successfuly found QR code
      summary: Gets Qr Code image and base Data
      tags:
      - qr
  /code/qr/deactivate/{serialNumber}:
    post:
      operationId: deactivateQRCode
      parameters:
      - explode: false
        in: path
        name: serialNumber
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: Successfully deactivate QR Code
      summary: Deactivate QR Code Code
components:
  schemas:
    QRCodeBaseData:
      example:
        image: ""
        serialNumber: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        url: url
      properties:
        serialNumber:
          format: uuid
          type: string
        url:
          type: string
        image:
          format: binary
          type: string
      required:
      - image
      - serialNumber
      - url
      type: object
    QRCodeCreateData:
      example:
        styleData:
          brightColor: brightColor
          fileName: QRcode.png
          margin: 6
          darkColor: darkColor
          size: 0
          marginColor: marginColor
        url: url
      properties:
        url:
          type: string
        styleData:
          $ref: '#/components/schemas/QRCodeStyleData'
      required:
      - url
      type: object
    QRCodeStyleData:
      example:
        brightColor: brightColor
        fileName: QRcode.png
        margin: 6
        darkColor: darkColor
        size: 0
        marginColor: marginColor
      properties:
        fileName:
          default: QRcode.png
          description: Name of QR Code image file
          type: string
        size:
          default: 25
          description: Size of QR Code in Pixel
          type: integer
        margin:
          default: 0
          description: Amount of space in pixels to add as a margin around the rendered
            QR Code
          type: integer
        brightColor:
          description: Color to be used for the "bright" parts of the QR Code. As
            HEX string
          type: string
        darkColor:
          description: Color to be used for the "dark" parts of the QR Code. As HEX
            string
          type: string
        marginColor:
          description: Color to be used for the "margin" part of the QR Code. As HEX
            string
          type: string
      type: object